Use own resolver for SitePreview
